Enable log sampling
Zapwriter learned about three new config options

    Config.SampleTick
    Config.SampleInitial
    Config.SampleThereafter

that, if defined, are passed to zapcore.NewSampler. This lets users
control how much gets written by each logger.

We want this because we can use it in go-carbon. When we rate-limit the
creation of new metrics, we log a helpful message with the metric name.
We also send our logs to an Elastic Search cluster for easy aggregation.
It turns out that logging every error message when rate-limiting abusive
creators can murder our ES cluster.